[WIP]Amends CSV file test
I'm trying to write tests to a) check we can receive CSV file types and run the rate method on them and b) raise an error for non JSON, non CSV file types.
I think it might be better for me to change the CurrencyExchange.rate method to accept a file type and make the changes in the CurrencyExchange class itself, rather than creating a new class. That's what I need to do next.