[WIP] refactors the rate method
We have begun refactoring the code so that it is suitable to accept CSV files. We will put all of the logic of locating the to and from currencies into seperate classes for JSON and CSV files. The rate method will receive the right floats to make this simple calculation.